src/requirements.txt
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiohttp-cors==0.7.0
aiohttp-retry==2.9.1
aiosignal==1.3.1
alabaster==1.0.0
albucore==0.0.24
albumentations==2.0.6
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
app-model==0.3.1
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
asyncssh==2.18.0
atpublic==5.0
attrs==24.2.0
autopage==0.5.2
babel==2.16.0
beautifulsoup4==4.12.3
billiard==4.2.1
bitsandbytes==0.42.0
bleach==6.2.0
build==1.2.2.post1
cachetools==5.5.0
cachey==0.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cliff==4.8.0
cloudpickle==3.1.0
cmaes==0.11.1
cmd2==2.5.7
colorama==0.4.6
coloredlogs==15.0.1
colorful==0.5.6
colorlog==6.9.0
comm==0.2.2
configobj==5.0.9
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dask==2024.12.1
datasets==3.1.0
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
dictances==1.5.3
dictdiffer==0.9.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
docutils==0.21.2
dpath==2.2.0
dulwich==0.22.6
dvc==3.58.0
dvc-data==3.16.7
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
efficientnet_pytorch==0.7.1
einops==0.8.1
entrypoints==0.4
et_xmlfile==2.0.0
evaluate==0.4.3
executing==2.1.0
fastapi==0.115.5
fastjsonschema==2.21.1
filelock==3.16.1
flatbuffers==25.2.10
flatten-dict==0.4.2
flexcache==0.3
flexparser==0.4
flufl.lock==8.1.0
fonttools==4.55.0
fqdn==1.5.1
freetype-py==2.5.1
frozenlist==1.5.0
fsspec==2024.9.0
funcy==2.0
gitdb==4.0.11
GitPython==3.1.43
gnureadline==8.2.13
google-api-core==2.23.0
google-auth==2.36.0
googleapis-common-protos==1.66.0
grandalf==0.8
grpcio==1.68.1
gto==1.7.2
h11==0.14.0
h5py==3.12.1
hausdorff==0.2.6
HeapDict==1.0.1
hsluv==5.0.4
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
huggingface-hub==0.26.3
humanfriendly==10.0
humanize==4.11.0
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
identify==2.6.3
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.6.0
imagesize==1.4.1
importlib_resources==6.4.5
in-n-out==0.2.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.30.0
ipywidgets==8.1.5
isoduration==20.11.0
iterative-telemetry==0.0.9
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.10.0
jsonargparse==4.35.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
kombu==5.4.2
lazy_loader==0.4
lightning==2.5.1.post0
lightning-utilities==0.11.9
linkify-it-py==2.0.3
llvmlite==0.43.0
locket==1.0.0
magicgui==0.10.0
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.14.0
miseval==1.3.0
mistune==3.0.2
moviepy==2.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
munch==4.0.0
napari==0.5.6
napari-console==0.1.3
napari-plugin-engine==0.2.0
napari-svg==0.2.1
natsort==8.4.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nodeenv==1.9.1
notebook==7.2.2
notebook_shim==0.2.4
npe2==0.7.8
numba==0.60.0
numpy==2.0.2
numpydoc==1.8.0
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.22.1
opencensus==0.11.4
opencensus-context==0.1.3
opencv-contrib-python==4.10.0.84
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
optuna==2.10.1
orjson==3.10.12
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
pbr==6.1.0
pexpect==4.9.0
pillow==10.4.0
Pint==0.24.4
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
pre_commit==4.0.1
pretrainedmodels==0.7.4
prettytable==3.12.0
proglog==0.1.11
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.0
psutil==6.1.0
psygnal==0.12.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-spy==0.4.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyconify==0.2.1
pycparser==2.22
pydantic==2.10.2
pydantic-compat==0.1.2
pydantic_core==2.27.1
pydot==3.0.3
pygit2==1.16.0
Pygments==2.18.0
pygtrie==2.5.0
PyOpenGL==3.1.9
pyparsing==3.2.0
pyperclip==1.9.0
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytorch-lightning==2.5.0.post0
pytz==2024.2
pyvista==0.44.2
PyYAML==6.0.2
pyzmq==26.2.0
qtconsole==5.6.1
QtPy==2.4.3
ray==2.39.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
roman-numerals-py==3.1.0
rootutils==1.0.7
rpds-py==0.22.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scikit-video==1.1.11
scipy==1.14.1
scmrepo==3.3.9
scooby==0.10.0
seaborn==0.13.2
segmentation_models_pytorch==0.5.0
semver==3.0.2
Send2Trash==1.8.3
sentry-sdk==2.19.2
setproctitle==1.3.4
setuptools==75.1.0
shellingham==1.5.4
shiboken6==6.9.1
shortuuid==1.0.13
shtab==1.7.1
SimpleITK==2.4.0
simsimd==6.2.1
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.36
sqltrie==0.11.1
stack-data==0.6.3
starlette==0.41.3
stevedore==5.4.0
stringzilla==3.11.3
superqt==0.7.1
sympy==1.13.1
tabulate==0.9.0
tensorboardX==2.6.2.2
terminado==0.18.1
textual==0.88.1
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==0.9.7
tinycss2==1.4.0
tokenizers==0.20.3
tomli_w==1.2.0
tomlkit==0.13.2
toolz==1.0.0
torch==2.5.1
torchio==0.20.8
torchmetrics==1.6.0
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
typer==0.14.0
types-python-dateutil==2.9.0.20241003
typeshed_client==2.7.0
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
vine==5.1.0
virtualenv==20.28.0
vispy==0.14.3
voluptuous==0.15.2
vtk==9.3.1
wandb==0.19.1
watchfiles==1.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
wheel==0.44.0
widgetsnbextension==4.0.13
wrapt==1.17.0
xgboost==2.1.3
xxhash==3.5.0
yarl==1.18.3
zc.lockfile==3.0.post1
